Apply more format changes
Fixed some issues with the flexbox positioning on the quote section,
also added text to the "cards" section and modified the order of the
flexboxes (added one class named card-sheet), imported more styles for
the fonts, added the sign up button to the hero section and made the
image placeholder. Added the anchor to the header links.